test(timestamp): assert real epoch-ns + sample-date event_date (failing)
Add tests that fail against the current seconds-vs-nanoseconds bug, where
deserialize.go stamps float64(UnixNano())/1e9 (fractional epoch SECONDS,
~1.75e9) onto a field named/typed as nanoseconds, collapsing every derived
event_date to 1970-01-01.

- pkg/xtcp/deserialize_test.go: new TestDeserialize_stampsTimestampNs drives a
  known poll time through the real Deserialize seam and asserts the produced
  record carries exact epoch nanoseconds (>= 1e18, == UnixNano incl. sub-second)
  and a same-day derived date. RED now: value is ~1.75e9, date is 1970-01-01.
- pkg/xtcp/destinations_s3parquet_schema_test.go: expect timestamp_ns physical
  kind INT64 (was DOUBLE). RED now.
- nix/microvms/self-test.nix: strengthen check 14b (s3parquet) to reject any
  1970-01-01 / stale event_date and any seconds-magnitude timestamp_ns, and add
  a plausible-range + sub-second guard to check 11 (ClickHouse).

+++ pkg/xtcp/destinations_s3parquet_schema_test.go
@@ -132,7 +132,7 @@ func TestS3ParquetSchema_columnTypes(t *testing.T) {
 		col      string
 		wantKind parquet.Kind
 	}{
-		{"timestamp_ns", parquet.Double},
+		{"timestamp_ns", parquet.Int64},
 		{"event_date", parquet.ByteArray},
 		{"hostname", parquet.ByteArray},
 		{"netns", parquet.ByteArray},
